requirements.txt
anthropic
pyautogui
pillow
python-dotenv
